+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Decoupled net power update period from charge current update period to provide more data points for power allocation moving average calculation. Please set "[Current update period](https://github.com/flashg1/SolarCharger/wiki/Installation#current-update-period)" in global defaults.
 - Reminder: Please [delete then re-add the integration](https://github.com/flashg1/SolarCharger/wiki/Configuration#how-to-delete-and-re-add-the-solarcharger-integration) if upgrading from v0.5.4 or prior.
 ### Improvement
+- Adjust charge current on synchronised update cycle across all chargers.
 - Added missing entities for config debug logging during startup.
 
 ## v0.6.1 2026-04-26